Move __iterators_possibly_equal to utils and use to check in-place for simd scan
