+++ include/oneapi/dpl/pstl/hetero/numeric_impl_hetero.h
@@ -18,6 +18,7 @@
 
 #include <iterator>
 #include "../parallel_backend.h"
+#include "../utils.h"
 #if _ONEDPL_BACKEND_SYCL
 #    include "dpcpp/execution_sycl_defs.h"
 #    include "algorithm_impl_hetero.h" // to use __pattern_walk2_brick
@@ -98,61 +99,6 @@ __pattern_transform_reduce(_ExecutionPolicy&& __exec, _ForwardIterator __first, 
 //------------------------------------------------------------------------
 template <typename T>
 struct ExecutionPolicyWrapper;
-
-// TODO In C++20 we may try to use std::equality_comparable
-template <typename _Iterator1, typename _Iterator2, typename = void>
-struct __is_equality_comparable : std::false_type
-{
-};
-
-// All with implemented operator ==
-template <typename _Iterator1, typename _Iterator2>
-struct __is_equality_comparable<
-    _Iterator1, _Iterator2,
-    std::void_t<decltype(::std::declval<::std::decay_t<_Iterator1>>() == ::std::declval<::std::decay_t<_Iterator2>>())>>
-    : std::true_type
-{
-};
-
-#if _ONEDPL_BACKEND_SYCL
-template <sycl::access::mode _Mode1, sycl::access::mode _Mode2, typename _T, typename _Allocator>
-bool
-__iterators_possibly_equal(const sycl_iterator<_Mode1, _T, _Allocator>& __it1,
-                           const sycl_iterator<_Mode2, _T, _Allocator>& __it2)
-{
-    const auto buf1 = __it1.get_buffer();
-    const auto buf2 = __it2.get_buffer();
-
-    // If two different sycl iterators belongs to the different sycl buffers, they are different
-    if (buf1 != buf2)
-        return false;
-
-    // We are unable to compare two sycl_iterator's if one of them is sub_buffer and assume that
-    // two different sycl iterators are equal.
-    if (buf1.is_sub_buffer() || buf2.is_sub_buffer())
-        return true;
-
-    return __it1 == __it2;
-}
-#endif // _ONEDPL_BACKEND_SYCL
-
-template <typename _Iterator1, typename _Iterator2>
-constexpr bool
-__iterators_possibly_equal(_Iterator1 __it1, _Iterator2 __it2)
-{
-    if constexpr (__is_equality_comparable<_Iterator1, _Iterator2>::value)
-    {
-        return __it1 == __it2;
-    }
-    else if constexpr (__is_equality_comparable<_Iterator2, _Iterator1>::value)
-    {
-        return __it2 == __it1;
-    }
-    else
-    {
-        return false;
-    }
-}
 
 template <typename _ExecutionPolicy, typename _Iterator1, typename _Iterator2, typename _UnaryOperation,
           typename _InitType, typename _BinaryOperation, typename _Inclusive>
